NAS-123271: Move functions from webworker to utils
